requirements.txt
# === Web Frameworks ===
Flask==3.1.1
flask-cors==6.0.1
Werkzeug==3.1.3
waitress==3.0.0
python-dotenv==1.1.0

# === Machine Learning / AI ===
tensorflow==2.19.0
numpy==2.1.0
pandas==2.2.2
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2

# === Database / Networking ===
pymongo==4.8.0
requests==2.32.3
